+++ notebooks/disable_ops.py
@@ -16,7 +16,7 @@
 import torch
 
 # %%
-from aioway.modes import PrintTorchDisp, PrintTorchFunc, mode_off
+from aioway.data import PrintTorchDisp, PrintTorchFunc, mode_off
 
 # %%
 a = torch.randn(3, 4)
