+++ notebooks/print_torch_ops.py
@@ -16,7 +16,7 @@
 import torch
 
 # %%
-from aioway.modes import PrintTorchDisp, PrintTorchFunc, fake_mode
+from aioway.data import PrintTorchDisp, PrintTorchFunc, fake_mode
 
 # %%
 a = torch.randn(3, 4)
