+++ src/aioway/_utils/decomps.py
@@ -65,7 +65,7 @@ def replace_tensors(
     Caused by `.device` / `.shape` / `.dtype` calls, which is used in `replace_tensors`.
     """
 
-    from aioway.modes import mode_off
+    from aioway.data import mode_off
 
     def maybe_replace(item):
         if not isinstance(item, torch.Tensor):
