+++ src/aioway/data/modes/README.md
