+++ src/aioway/data/modes/aten/README.md
