+++ src/aioway/data/modes/aten/aten.py
@@ -11,10 +11,10 @@ from collections import abc as cabc
 from torch import _ops
 
 from aioway._utils import camel_to_snake, find_nested_tensors, render_fcall
-from aioway.sessions import Cost
+from aioway.data.sessions import Cost
 
 if typing.TYPE_CHECKING:
-    from aioway.modes import TorchDispThunk
+    from aioway.data import TorchDispThunk
 
 __all__ = ["Aten", "aten_dcls", "find_aten", "all_aten_overrides"]
