+++ src/aioway/data/modes/aten/fn.py
@@ -10,7 +10,7 @@ from .aten import Aten, find_aten
 from .guards import is_aten_op
 
 if typing.TYPE_CHECKING:
-    from aioway.modes import TorchDispThunk
+    from aioway.data import TorchDispThunk
 
 __all__ = ["AtenThunk"]
