+++ src/aioway/data/modes/aten/getitem.py
@@ -6,7 +6,7 @@ import typing
 import torch
 from torch import ops
 
-from aioway.sessions import Cost
+from aioway.data.sessions import Cost
 
 from .aten import Aten, aten_dcls
