+++ src/aioway/data/schemas/README.md
