+++ src/aioway/data/schemas/__init__.py
