+++ src/aioway/data/schemas/_bases.py
