+++ src/aioway/data/schemas/attrs.py
@@ -129,7 +129,7 @@ class Attr:
         Generate a random tensor. This should be used under fake mode.
         """
 
-        from aioway.modes import fake_mode
+        from aioway.data import fake_mode
 
         with fake_mode():
             return torch.zeros(
