+++ src/aioway/data/schemas/devices.py
