+++ src/aioway/data/schemas/shapes.py
