+++ src/aioway/data/schemas/utils.py
