+++ src/aioway/nn/__init__.py
@@ -2,5 +2,6 @@
 
 from .deductions import *
 from .emits import *
+from .rebuild import *
 from .rewrites import *
 from .tspecs import *
